+++ packages/storybook-dark-mode/src/components/dark_mode_toggle.tsx
@@ -2,9 +2,9 @@ import { MoonIcon, SunIcon } from '@storybook/icons'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { Button } from 'storybook/internal/components'
 import { type API, addons, useParameter } from 'storybook/manager-api'
-import { mergeThemeWithBrandConfig } from '../internal/utils/merge_theme_with_brand_config'
-import type { DarkModeStore } from '../internal/utils/store'
-import { prefersDark, store, updateManager, updatePreview, updateStore } from '../internal/utils/store'
+import { mergeThemeWithBrandConfig } from '../_internal/utils/merge_theme_with_brand_config'
+import type { DarkModeStore } from '../_internal/utils/store'
+import { prefersDark, store, updateManager, updatePreview, updateStore } from '../_internal/utils/store'
 import { DARK_MODE_EVENT_NAME, UPDATE_DARK_MODE_EVENT_NAME } from '../static/constants'
 import type { Mode } from '../static/types'
